move database handlers
