+++ src/lib/database/handleConnection.ts
