+++ src/lib/database/handleGuildCheck.ts
