+++ src/lib/database/handleRegisterGuild.ts
