+++ src/lib/database/handleRegisterGuildEntity.ts
